# api-specification/target/classes/openapi.yaml
openapi: 3.1.0
info:
  title: Healfiness Webshop API
  description: Enterprise REST API for Healfiness Webshop
  version: 1.0.0
servers:
  - url: 'https://localhost:8080/api/v1'
    description: Local development server
  - url: https://api.healfiness.com/api/v1
    description: Production
...
security:
  - bearerAuth: []
  - basicAuth: []
paths:
...
  /addresses:
    get:
      tags:
        - Addresses
      summary: Get addresses
      description: Retrieve a list of all addresses
      operationId: getAddresses
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: A list of addresses
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressPageResponse'
    post:
      tags:
        - Addresses
      summary: Create an address
      description: Create a new address with the provided details
      operationId: createAddress
      requestBody:
        required: true
        description: The address details to create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressCreateRequest'
      responses:
        '201':
          description: Address created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressResponse'
        '400':
          description: Bad request - invalid address details supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
  /orders:
    get:
      tags:
        - Orders
      summary: Get orders
      description: Retrieve a list of all orders
      operationId: getOrders
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: A list of orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPageResponse'
    post:
      tags:
        - Orders
      summary: Create an order
      description: Create a new order with the provided details
      operationId: createOrder
      requestBody:
        required: true
        description: The order details to create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Bad request - invalid order details supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /orders/{orderId}:
    get:
      tags:
        - Orders
      summary: Find order by identifier
      description: Retrieve details of a specific order by its identifier
      operationId: findOrderById
      parameters:
        - name: orderId
          in: path
          required: true
          description: The ID of the order to fetch
          schema:
            $ref: '#/components/schemas/Id'
      responses:
        '200':
          description: Order details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Bad request - invalid order ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    patch:
      tags:
        - Orders
      summary: Update an order partially by identifier
      description: Update a specific order partially by its identifier
      operationId: updateOrder
      parameters:
        - name: orderId
          in: path
          required: true
          description: The ID of the order to update
          schema:
            $ref: '#/components/schemas/Id'
      requestBody:
        required: true
        description: The updated order details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdateRequest'
      responses:
        '200':
          description: Order updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Bad request - invalid order details supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

    basicAuth:
      type: http
      scheme: basic
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Sort:
      name: sort
      in: query
      required: false
      schema:
        type: array
        items:
          type: string
          example: createdBy,asc
...
  schemas:
...
    Id:
      type: integer
      format: int64
      readOnly: true
      example: 123
    Money:
      type: number
      format: double
      example: 29.99
    Timestamp:
      type: string
      format: date-time
      example: "2026-05-07T14:30:00Z"
    UserName:
      type: string
      minLength: 3
      maxLength: 15
      example: "johndoe"
...
    PageMetaData:
      type: object
      properties:
        page:
          type: integer
          example: 0

        size:
          type: integer
          example: 20

        totalElements:
          type: integer
          example: 1

        totalPages:
          type: integer
          example: 1

        last:
          type: boolean
          example: true

        sortOrder:
          $ref: '#/components/schemas/SortOrder'
    PageResponse:
      type: object
      required:
        - content
        - pageMetaData
      properties:
        content:
          type: array
          items: { }

        pageMetaData:
          $ref: '#/components/schemas/PageMetaData'
    ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
        - detail
        - timestamp
      properties:
        type:
          type: string
          example: "https://api.healfiness.com/errors/validation"

        title:
          type: string
          example: "Validation Error"

        status:
          type: integer
          example: 400

        detail:
          type: string
          example: "Validation failed for request body."

        timestamp:
          $ref: '#/components/schemas/Timestamp'

        violations:
          type: array
          items:
            $ref: '#/components/schemas/ValidationViolation'
    ValidationViolation:
      type: object
      properties:
        field:
          type: string
          example: "email"

        message:
          type: string
          example: "must be a valid email"
    MetaData:
      type: object
      properties:
        createUser:
          $ref: '#/components/schemas/UserName'

        createDate:
          $ref: '#/components/schemas/Timestamp'

        lastModifiedUser:
          $ref: '#/components/schemas/UserName'

        lastModifiedDate:
          $ref: '#/components/schemas/Timestamp'

        version:
          type: integer
          format: int64
          example: 1
    SortOrder:
      type: object
      properties:
        property:
          type: string
          example: "createUser"

        direction:
          type: string
          enum: [ASC, DESC]

    # =========================================================
    # AUTH
    # =========================================================
...
    OrderStatus:
      type: string
      enum:
        - PENDING
        - AWAITING_PAYMENT
        - PAID
        - PROCESSING
        - ON_HOLD
        - BACK_ORDER
        - PARTIALLY_SHIPPED
        - SHIPPED
        - DELIVERED
        - COMPLETED
        - CANCELLED
        - REFUNDED
        - RETURNED
        - FAILED
    AddressType:
      type: string
      enum:
        - HOME
        - DELIVERY
        - PICKUP
        - INVOICE
...
    ProductSummary:
      type: object
      required:
        - id
        - name
        - price
      properties:
        id:
          $ref: '#/components/schemas/Id'

        name:
          type: string
          minLength: 2
          maxLength: 150
          example: "Protein Powder"

        price:
          $ref: '#/components/schemas/Money'
...
    IsoCountryCodeSummary:
      type: object
      required:
        - isoCountryCodesId
        - isoCountryCode
        - countryName
      properties:
        isoCountryCodesId:
          $ref: '#/components/schemas/Id'

        isoCountryCode:
          type: string
          minLength: 2
          maxLength: 3
          example: "DE"

        countryName:
          type: string
          example: "Germany"
    IsoCountryCodeResponse:
      allOf:
        - $ref: '#/components/schemas/IsoCountryCodeSummary'
        - type: object
          properties:
            abstractObjectMetaData:
              $ref: '#/components/schemas/MetaData'
...
    CitySummary:
      type: object
      required:
        - citiesId
        - shortName
        - name
      properties:
        citiesId:
          $ref: '#/components/schemas/Id'

        shortName:
          type: string
          minLength: 2
          maxLength: 3
          example: "HAM"

        name:
          type: string
          example: "Hamburg"
    CityResponse:
      allOf:
        - $ref: '#/components/schemas/CitySummary'
        - type: object
          properties:
            isoCountryCode:
              $ref: '#/components/schemas/IsoCountryCodeResponse'

            abstractObjectMetaData:
              $ref: '#/components/schemas/MetaData'
...
    CityCreateRequest:
      type: object
      required:
        - name
        - isoCountryCodeId
      properties:
        name:
          type: string

        shortName:
          type: string
          minLength: 2
          maxLength: 5

        isoCountryCodeId:
          $ref: '#/components/schemas/Id'

    # =========================================================
    # ADDRESS
    # =========================================================
    AddressResponse:
      type: object
      required:
        - id
        - street
        - houseNumber
        - postalCode
        - city
        - type
        - userId
      properties:
        id:
          $ref: '#/components/schemas/Id'

        street:
          type: string
          maxLength: 255
          example: "Main Street"

        houseNumber:
          type: string
          maxLength: 20
          example: "42A"

        postalCode:
          type: string
          maxLength: 20
          example: "20095"

        suffix:
          type: string
          maxLength: 20
          nullable: true

        type:
          $ref: '#/components/schemas/AddressType'

        city:
          $ref: '#/components/schemas/CityResponse'

        userId:
          $ref: '#/components/schemas/Id'

        abstractObjectMetaData:
          $ref: '#/components/schemas/MetaData'
    AddressPageResponse:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/AddressResponse'
    AddressCreateRequest:
      type: object
      required:
        - street
        - houseNumber
        - postalCode
        - city
        - type
        - userId
      properties:
        street:
          type: string

        houseNumber:
          type: string

        postalCode:
          type: string

        suffix:
          type: string
          nullable: true

        city:
          $ref: '#/components/schemas/CityCreateRequest'

        type:
          $ref: '#/components/schemas/AddressType'

        userId:
          $ref: '#/components/schemas/Id'
...
    OrderItemResponse:
      type: object
      required:
        - id
        - quantity
        - product
        - singlePriceAtOrderTime
      properties:
        id:
          $ref: '#/components/schemas/Id'

        quantity:
          type: integer
          minimum: 1
          example: 1

        singlePriceAtOrderTime:
          $ref: '#/components/schemas/Money'

        product:
          $ref: '#/components/schemas/ProductSummary'

        abstractObjectMetaData:
          $ref: '#/components/schemas/MetaData'
    OrderItemCreateRequest:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          $ref: '#/components/schemas/Id'

        quantity:
          type: integer
          minimum: 1

    # =========================================================
    # ORDER
    # =========================================================
    OrderSummary:
      properties:
        id:
          $ref: '#/components/schemas/Id'

        orderStatus:
          $ref: '#/components/schemas/OrderStatus'

        totalAmount:
          $ref: '#/components/schemas/Money'
    OrderResponse:
      allOf:
        - $ref: '#/components/schemas/OrderSummary'
        - type: object
          properties:
            orderDate:
              $ref: '#/components/schemas/Timestamp'

            totalQuantity:
              type: integer
              example: 1

            orderItems:
              type: array
              items:
                $ref: '#/components/schemas/OrderItemResponse'

            userId:
              $ref: '#/components/schemas/Id'

            abstractObjectMetaData:
              $ref: '#/components/schemas/MetaData'
    OrderPageResponse:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/OrderResponse'
    OrderCreateRequest:
      type: object
      required:
        - userId
        - orderItems
        - orderStatus
        - quantity
        - totalAmount
      properties:
        userId:
          $ref: '#/components/schemas/Id'

        orderItems:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItemCreateRequest'

        orderStatus:
          $ref: '#/components/schemas/OrderStatus'

        quantity:
          type: integer

        totalAmount:
          $ref: '#/components/schemas/Money'
    OrderUpdateRequest:
      type: object
      properties:
        orderStatus:
          $ref: '#/components/schemas/OrderStatus'

        orderItems:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItemCreateRequest'

        quantity:
          type: integer

        totalAmount:
          $ref: '#/components/schemas/Money'

    # =========================================================
    # SHOPPING CART
    # =========================================================
...